Fix monochrome aesthetic violations and move ChatSessionOrchestrator
- Remove heavy pill background from CollapsibleButton (rounded-lg, bg-surface-tertiary)
  in favor of minimal text+chevron styling
- Neutralize light-mode border colors from slate (blue-gray) to pure gray
- Replace blue glow shadows with neutral black
- Fix ToolPermissionInline dark icon wrapper opacity (white/10 → white/5)
- Move ChatSessionOrchestrator into chat-window/ where it belongs
